Refactor listener facade into separate module to avoid LOGGER import cycle
Move listener facade/adaptation logic out of robot.output.listeners into a new
robot.output.listenerfacade module so that the facade can be reused without
importing the global LOGGER (and thus without creating circular imports with
robot.output.logger).

Key changes:
- Add src/robot/output/listenerfacade.py:
  - Introduce ListenerFacade.from_object() to wrap an already-instantiated
    listener object into the correct v2/v3 facade based on
    ROBOT_LISTENER_API_VERSION.
  - Keep listener priority parsing and API version validation behavior.
  - Implement ListenerMethod to invoke underlying listener methods and to report
    failures via injected callbacks (error/info), removing the hard dependency
    on the global LOGGER from the facade layer.
  - Preserve v2/v3-specific attribute mapping and body item/type conversions
    (e.g. keyword start/end, v2/v3 argument/attrs handling).

- Update src/robot/output/listeners.py:
  - Remove the facade implementation from this module and keep it focused on
    listener management and string-spec importing.
  - Add create_listener_facade() that:
    - imports listener objects from string specifications using Importer()
      with LOGGER (existing behavior),
    - wraps imported objects via ListenerFacade.from_object(),
    - injects LOGGER.error/LOGGER.info into the facade for runtime reporting,
    - preserves “Taking listener '<name>' into use failed: …” DataError
      wrapping semantics, using the original string spec as the name when
      applicable.

- Update utest/output/test_listeners.py:
  - Import ListenerFacade from robot.output.listenerfacade (new location).
  - Use ListenerFacade.from_object() directly in the priority tests, supplying
    dummy error/info callbacks.
  - Make the test listener string spec derive from __name__
    (f"{__name__}.ListenAll") so the listener can be imported both when the
    module is executed via unittest (package import) and via the custom utest
    runner (top-level module import).
